moved lambda desugaring to desugaring step
