+++ src/compiler/mod.rs
@@ -20,6 +20,7 @@ pub mod gen;
 
 pub mod token;
 pub mod ast; // high level pre-macro IR
+pub mod rule; // macro transformation
 pub mod cst; // post-macro IR
 
 pub mod syntax;
